requirements.txt
pandas
python-dotenv
langchain-google-genai
langchain-core
requests